==> Part 1 - Aggregate Service Incidents, Interactions, Cases and Install Base.ipynb <==
# import relevant libraries
import pandas as pd
import scipy.stats
import numpy as np
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('float_format', '{:,.2f}'.format)

# Importing datasets

# Reading data tables
cases = pd.read_csv('Cases.csv')
interactions = pd.read_csv('Interactions.csv')
installbase = pd.read_csv('Install_Base.csv')
service = pd.read_csv('Incidents.csv')

# Creating aggregated features

Before join with revenue table, we created aggregated features for Service Incidents, Cases, Interactions and Install Base table.

## Service Incidents

service.head()

service.shape

### Preprocessing before aggregating

# check duplicates
service.drop_duplicates(inplace = True)
service.shape

# check column names and null values
service.info()

# drop unnecessary columns
service.drop(['DUNS_NUMBER','SFDC_ID','GROUP_NAME'],axis = 1, inplace = True)

service.head()

### Aggregating

# aggregate NO_OF_SVC_INCIDENTS, NO_OF_REPEAT_CALLS_30D, NO_OF_FTF_CALLS_30D, LAST_SERVICE_EVENT_DATE
service = service.groupby('CUSTOMER_SITE_ID').agg({'NO_OF_SVC_INCIDENTS': 'sum',
                                                   'NO_OF_REPEAT_CALLS_30D': 'sum',
                                                   'NO_OF_FTF_CALLS_30D': 'sum'}).reset_index()
service.head()

service.shape

# rename columns
service = service.rename(columns = {
                                    "NO_OF_SVC_INCIDENTS": "Total_SVC_Incidents",
                                    "NO_OF_REPEAT_CALLS_30D": "Total_Repeat_Calls",
                                    "NO_OF_FTF_CALLS_30D": "Total_FTF_Calls"})

service.head()

# check all customer sites are unique
len(service['CUSTOMER_SITE_ID'])/len(service['CUSTOMER_SITE_ID'].unique())

## Interactions

interactions.head()

interactions.shape

### Preprocessing before aggregating

# check duplicates
interactions.drop_duplicates(inplace = True)
interactions.shape

# check column names and null values
interactions.info()

# drop unnecessary columns
interactions.drop('Interaction_Amount',axis = 1,inplace = True)

# fill in null values in Type
interactions['Type'] =  interactions['Type'].fillna('Other')

# drop rows with null values
interactions = interactions.dropna()

interactions.info()

interactions.head()

# drop unnecessary columns
interactions.drop(['Task_Subtype','Last_Modified_Date','SFDC_ID','Assigned','Created_Date'],axis = 1, inplace = True)

interactions.head()

### Aggregating

# aggregate Type
interactions1 = interactions[['CUSTOMER_SITE_ID','Type']]
interactions2 = interactions1.groupby('CUSTOMER_SITE_ID')['Type'].apply(lambda x: x.mode()[0]).reset_index()
interactions2.head()

# # check all customer sites are unique
len(interactions2['CUSTOMER_SITE_ID'])/len(interactions2['CUSTOMER_SITE_ID'].unique())

# aggregate Visit and Task
interactions = interactions.drop('Type',axis = 1)
interactions3 = interactions.groupby('CUSTOMER_SITE_ID')[['Visit','Task']].sum().reset_index()
interactions3.head()

# combine aggregated variables
interactions = pd.merge(interactions2, interactions3, how = 'inner')
interactions.head()

# rename columns
interactions = interactions.rename(columns = {"Type": 'Most_Frequent_Interaction_Type',
                                              "Visit": "Total_Visits",
                                              "Task": "Total_Tasks"})
interactions.head()

# # check all customer sites are unique
len(interactions['CUSTOMER_SITE_ID'])/len(interactions['CUSTOMER_SITE_ID'].unique())

## Cases

cases.sort_values(by='NO_OF_CASES', ascending=False).head(10)

cases.shape

### Preprocessing before aggregating

# check duplicates
cases.drop_duplicates(inplace = True)
cases.shape

# check column names and null values
cases.info()

# drop rows with null values in CUSTOMER_SITE_ID, CASE_ORIGIN
cases = cases.dropna(axis = 0, subset = ['CUSTOMER_SITE_ID'])
cases = cases.dropna(axis = 0, subset = ['CASE_ORIGIN'])
cases.info()

# drop unnecessary columns
cases.drop(['SFDC_ID','DUNS_NUMBER','MAX_MODIFIED_DATE','CX_CASE_L1_REASON','CX_CASE_L2_REASON'],
            axis = 1, inplace = True)
cases.info()

cases.head()

### Aggregating

# aggregate CASE_ORIGIN
cases01 = cases.groupby(['CUSTOMER_SITE_ID','CASE_ORIGIN'])['NO_OF_CASES'].sum().reset_index()
cases02 = cases01.groupby('CUSTOMER_SITE_ID')['NO_OF_CASES'].max().reset_index()
cases03 = cases02.merge(cases01, how = 'inner', on=['CUSTOMER_SITE_ID','NO_OF_CASES'])
cases03 = cases03.groupby('CUSTOMER_SITE_ID')['CASE_ORIGIN'].apply(lambda x: x.mode()[0]).reset_index()
cases03.head(10)

len(cases03.CUSTOMER_SITE_ID.value_counts())/len(cases03)

# aggregate CASE_REASON
cases11 = cases.groupby(['CUSTOMER_SITE_ID','CASE_REASON'])['NO_OF_CASES'].sum().reset_index()
cases12 = cases11.groupby('CUSTOMER_SITE_ID')['NO_OF_CASES'].max().reset_index()
cases13 = cases12.merge(cases11, how = 'inner', on=['CUSTOMER_SITE_ID','NO_OF_CASES'])
cases13 = cases13.groupby('CUSTOMER_SITE_ID')['CASE_REASON'].apply(lambda x: x.mode()[0]).reset_index()
cases13.head(10)

len(cases13.CUSTOMER_SITE_ID.value_counts())/len(cases13)

# aggregate NO_OF_CASES
cases1 = cases[['CUSTOMER_SITE_ID','NO_OF_CASES']]
cases2 = cases1.groupby('CUSTOMER_SITE_ID')['NO_OF_CASES'].sum().reset_index()
cases = cases2
cases2.head()

# rename columns
cases = cases.rename(columns = {"NO_OF_CASES": "Total_Cases"})
cases.head()

cases = pd.merge(cases, cases03, on='CUSTOMER_SITE_ID').merge(cases13, on='CUSTOMER_SITE_ID')

cases = cases.rename(columns = {'CASE_ORIGIN':'Max_Case_Origin','CASE_REASON': 'Max_Case_Reason'})

cases.head()

# # check all customer sites are unique
len(cases['CUSTOMER_SITE_ID'])/len(cases['CUSTOMER_SITE_ID'].unique())

## Install Base

installbase.head()

installbase.shape

### Preprocessing before aggregating

# check duplicates
installbase.drop_duplicates(inplace = True)
installbase.shape

# check column names and null values
installbase.info()

# copy INSTALL_DATE for later steps
installbase['INSTALL_DATE1'] = installbase['INSTALL_DATE']

# convert date variables to datetime data type
installbase[["INSTALL_DATE","INSTALL_DATE1","CONTRACT_START_DATE", "CONTRACT_END_DATE","INSTANCE_LAST_UPDATE_DATE"]] = installbase[["INSTALL_DATE","INSTALL_DATE1","CONTRACT_START_DATE", "CONTRACT_END_DATE","INSTANCE_LAST_UPDATE_DATE"]].apply(pd.to_datetime)

# check data types
installbase.info()

# create a new column called Contract_length
installbase['Contract_length'] = (installbase['CONTRACT_END_DATE'] - installbase['CONTRACT_START_DATE']).dt.days

# fill missing value in this column using 0
installbase['Contract_length'] = installbase['Contract_length'].fillna(0)

# drop unnecessary columns
installbase.drop(['INSTANCE_ID','SFDC_ID','SITE_NUMBER','DUNS_NUMBER','PRODUCT_FAMILY','CONTRACT_START_DATE','CONTRACT_END_DATE'], 
                 axis = 1, inplace = True)
installbase.head()

# check column names and null values
installbase.info()

installbase['CONTRACT_FLAG'].value_counts()

# substitute Y/N with 1/0 in CONTRACT_FLAG
installbase['CONTRACT_FLAG'] = installbase['CONTRACT_FLAG'].apply(lambda x: 1 if x=='Y' else 0)

installbase.head()

# check column names and null values
installbase.info()

# drop rows with null values in STATUS, INSTALL_DATE
installbase = installbase.dropna(axis = 0, subset = ['STATUS'])
installbase = installbase.dropna(axis = 0, subset = ['INSTALL_DATE'])

# check column names and null values
installbase.info()

installbase.CONTRACT_CATEGORY.value_counts()

# fill in null values in CONTRACT_CATEGORY
installbase['CONTRACT_CATEGORY'] =  installbase['CONTRACT_CATEGORY'].fillna('No Contract')

# check column names and null values
installbase.info()

installbase.head()

### Aggregating

# copy columns to be aggregated
installbase2 = installbase[['INSTALL_DATE','INSTALL_DATE1','CONTRACT_FLAG','CUSTOMER_SITE_ID',
                            'INSTANCE_LAST_UPDATE_DATE','Contract_length', 'CONTRACT_CATEGORY']]

installbase2.head()

# count number of install bases
installbase3 = installbase[['CUSTOMER_SITE_ID']]
installbase3 = installbase3.groupby('CUSTOMER_SITE_ID').size().reset_index()
installbase3 = installbase3.rename(columns = {0: 'Num_of_Install_Bases'})
installbase3.head()

# aggregate CONTRACT_FLAG, Contract_length, INSTANCE_LAST_UPDATE_DATE, INSTALL_DATE

installbase2['Contract_length'] = installbase2['Contract_length'].replace(0, np.nan)
installbase2 = installbase2.groupby('CUSTOMER_SITE_ID').agg({'CONTRACT_FLAG': 'sum',
                                                             'Contract_length': 'mean',     
}).reset_index()
# rename columns
installbase2 = installbase2.rename(columns = {"CONTRACT_FLAG": 'Total_Contracts'})

installbase2['Contract_length'] = installbase2['Contract_length'].replace(np.nan,0)
installbase2.head()

# check all customer sites are unique
len(installbase2['CUSTOMER_SITE_ID'])/len(installbase2['CUSTOMER_SITE_ID'].unique())

# merge the above 2 tables
installbase4 = pd.merge(installbase2, installbase3, how = 'inner')
installbase4.head()

# aggregate STATUS
installbase['Num_of_Active_Install_Bases'] = installbase['STATUS'].map({'Inactive - Unconfirmed': 0,
                                                       'Inactive - Confirmed': 0,
                                                       'Active - Unconfirmed': 1,
                                                       'Active - Confirmed': 1})
installbase5 = installbase.groupby('CUSTOMER_SITE_ID')['Num_of_Active_Install_Bases'].sum().reset_index()
installbase5.head()

# aggregate CONTRACT_CATEGORY
installbase6 = installbase.groupby('CUSTOMER_SITE_ID')['CONTRACT_CATEGORY'].apply(lambda x: x.mode()[0]).reset_index()
installbase6 = installbase6.rename(columns = {"CONTRACT_CATEGORY": 'Contract_Category'})
installbase6.head()

# merge the above tables
installbase = pd.merge(installbase5, installbase4, how = 'inner').merge(installbase6, how = 'inner')
installbase.head()

installbase.info()

# calculate a column called Number_of_Inactive_Sites
installbase['Num_of_Inactive_Install_Bases'] = installbase['Num_of_Install_Bases'] - installbase['Num_of_Active_Install_Bases']

installbase.head()

# check all customer sites are unique
len(installbase['CUSTOMER_SITE_ID'])/len(installbase['CUSTOMER_SITE_ID'].unique())

# Exporting result tables

service.to_csv("service_aggregated.csv")
interactions.to_csv("interactions_aggregated.csv")
cases.to_csv("cases_aggregated.csv")
installbase.to_csv("installbase_aggregated.csv")
