==> Part 2 - Join All Tables.ipynb <==
# import relevant libraries
import pandas as pd
import scipy.stats
import numpy as np
import datetime as dt
import seaborn as sns
import matplotlib.pyplot as plt
%matplotlib inline

pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('float_format', '{:,.2f}'.format)

# Importing datasets

# import datasets
cases = pd.read_csv('cases_aggregated.csv',index_col = 0)
interactions = pd.read_csv('interactions_aggregated.csv',index_col = 0)
installbase = pd.read_csv('installbase_aggregated.csv',index_col = 0)
service = pd.read_csv('service_aggregated.csv',index_col = 0)
combined = pd.read_csv('Combined_Site_Rev_Prod_Price.csv')

# Checking the table combined in SQL

combined.head()

combined.shape

# check duplicates
combined.drop_duplicates(inplace = True)
combined.shape

# check column names
combined.info()

# drop unnecessary columns
combined.drop(['Site_ID', 'yYEAR', 'ITEM_ID.1', 'SFDC_ID.1', 'DUNS_NUMBER.1', 'ITEM_ID.2'],axis = 1, inplace = True)

# check column names
combined.info()

# Join all tables using left join

# merge all tables
df = pd.merge(combined, service, how = 'left',on = 'CUSTOMER_SITE_ID')
df = pd.merge(df, interactions, how = 'left',on = 'CUSTOMER_SITE_ID')
df = pd.merge(df, cases, how = 'left',on = 'CUSTOMER_SITE_ID')
df = pd.merge(df, installbase, how = 'left',on = 'CUSTOMER_SITE_ID')

df.head()

df.shape

# check column names and null values
df.isnull().sum()

# Exporting result table

df.info()

df.to_csv("Final_Merged_File.csv")
