==> Part 7 - Analysis Dataset Creation.ipynb <==
# import required libraries for dataframe and visualization
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import datetime as dt
from lifelines import KaplanMeierFitter

# Ensuring all rows and columns are visible
pd.set_option('display.max_columns', None)
pd.set_option('display.max_rows', None)
pd.set_option('float_format', '{:,.2f}'.format)

# Reading the data on which analysis needs to be done

df = pd.read_csv('cleaned_data.csv',index_col = 0)
df.head()

# set Transaction Date to datetime 
df['TRX_DATE'] = pd.to_datetime(df['TRX_DATE'])

df.shape

# read variables dataset
variables = pd.read_csv('variables_with_churn.csv',index_col = 0 )
variables.head()

variables.groupby('Churned_365')['Tenure'].mean()

variables.shape

df.shape

variables.head()

# merge into a big dataset
df_final = pd.merge(df, variables, on='CUSTOMER_SITE_ID')
df_final.head()

df_final.columns = df_final.columns.str.replace('_x','',regex = True)

df_final.head()

df_final.drop(['SHORT_VERTICAL_y',
       'POSTAL_CODE_y', 'CUSTOMER_CLASS_y', 'TERRITORY_TYPE_y',
       'SUPPLIES_SEGMENTATION_y', 'SUPPLIES_DECLINE_REASONS_y',
       'DUNS_NUMBER_y', 'Total_SVC_Incidents_y', 'Total_Repeat_Calls_y',
       'Total_FTF_Calls_y', 'Most_Frequent_Interaction_Type_y',
       'Total_Visits_y', 'Total_Cases_y',
       'Num_of_Active_Install_Bases_y', 'Total_Contracts_y',
       'Contract_length_y', 'Contract_Category_y',
       'STRATEGIC_ACCOUNTS_y', 'TERRITORY_REGION_y'], axis = 1,inplace = True)

df_final.head()

df_final.sort_values('CUSTOMER_SITE_ID',ascending = True, inplace=True)

df_final.reset_index(inplace=True)

df_final = df_final.iloc[:,1:]

df_final.head()

df_final.shape

df_final.to_csv('analysis.csv')
